requirements.txt
pypdf
requests
pydantic
python-dotenv
